Feature/front/proyectos banco

+++ src/app/servicios/proyectos.service.ts
@@ -1,6 +1,7 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
+import UserNameBean from '../beans/UserNameBean';
 
 @Injectable({
   providedIn: 'root'
@@ -9,13 +10,31 @@ export class ProyectosService {
 
   constructor(private httpClient: HttpClient) { }
   getDatos(url: string): Observable<any> {
-    return this.httpClient.get(url);
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.get(url, {headers:{'Authorization': sessionStorage['token']}});
   }
-  saveOrUpdate(url:string, objeto:Object):Observable<any>{
-    return this.httpClient.post(url,{body:objeto})
+  saveOrUpdate(url:string, objeto:any):Observable<any>{
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.post(url,objeto, {headers:{'Authorization': sessionStorage['token']}})
   }
-  delete(url:string):void{
-    this.httpClient.delete(url);
+  delete(url:string):Observable<any>{
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.delete(url, {headers:{'Authorization': sessionStorage['token']}});
   }
+  patch(url:string,objeto:any):Observable<any>{
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.patch(url,objeto, {headers:{'Authorization': sessionStorage['token']}});
+  }
+  getDatosSinHeader(url: string): Observable<any> {
+    return this.httpClient.get(url)
+  }
+  patchSinHeader(url:string,user:any):Observable<any>{
+    return this.httpClient.patch(url,user)
+  }
+  
 
 }
